+++ backend/src/main/java/org/ntnu/idi/idatt2105/project/entity/CompletedQuiz.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class CompletedQuiz {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int completedQuizId;
+    private Long completedQuizId;
 
     @Column(name = "score")
     private int score;
