+++ backend/src/main/java/org/ntnu/idi/idatt2105/project/entity/QuestionType.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class QuestionType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int typeId;
+    private Long typeId;
 
     @Column(name = "type_name")
     private String type;
